app/requirements.txt
# Chainlit Web App Dependencies

# Core dependencies
chainlit>=1.0.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# Optional (for better performance)
asyncio>=3.4.3
